requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.14
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.1
dataclasses-json==0.5.14
docx2txt==0.8
duckduckgo-search==3.9.3
EbookLib==0.18
effdet==0.4.1
embedchain==0.0.74
emoji==2.8.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.104.0
filelock==3.12.4
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
grpcio==1.59.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
huggingface-hub==0.17.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-resources==6.1.0
iopath==0.1.10
Jinja2==3.1.2
joblib==1.3.2
jq==1.6.0
kiwisolver==1.4.5
langchain==0.0.279
langdetect==1.0.9
langsmith==0.0.49
layoutparser==0.3.4
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.1
olefile==0.46
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.15.1
openai==0.27.10
opencv-python==4.8.1.78
openpyxl==3.1.2
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.1.0
portalocker==2.8.2
posthog==3.0.2
protobuf==4.24.4
pulsar-client==3.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pypandoc==1.12
pyparsing==3.1.1
pypdf==3.16.4
pypdfium2==4.22.0
PyPika==0.48.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.0.1
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.4.0
regex==2023.10.3
replicate==0.15.4
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.5
SQLAlchemy==2.0.22
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
timm==0.9.7
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.34.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
unstructured==0.10.24
unstructured-inference==0.7.7
unstructured.pytesseract==0.3.12
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.18.0
watchfiles==0.21.0
websockets==11.0.3
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.2
youtube-transcript-api==0.6.1
